lambda-updates/requirements-update.txt
# Add these dependencies to your existing Lambda function requirements.txt

PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0

# If you don't have these already:
boto3>=1.34.0
botocore>=1.34.0